requirements.txt
absl-py==2.0.0
accelerate==0.34.2
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attr==0.3.1
attrs==23.1.0
azure-core==1.31.0
azure-storage-blob==12.23.0
backcall==0.2.0
beartype==0.16.4
bert-score==0.3.13
bioc==2.0.post5
bleach==5.0.1
blobfile==2.1.0
boto==2.49.0
boto3==1.35.24
botocore==1.35.24
boxing==0.1.4
cachetools==5.3.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.1.0
cryptography==43.0.1
cycler==0.11.0
Cython==3.0.6
DALL-E==0.1
dataclasses-json==0.6.7
datasets==2.14.5
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dm-tree==0.1.8
docopt==0.6.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
einops==0.7.0
exceptiongroup==1.1.3
executing==0.8.3
expiringdict==1.2.2
face-alignment==1.4.1
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gast==0.5.4
gco-wrapper==3.0.9
gensim==4.3.3
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.1.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
GPUtil==1.4.0
greenlet==3.0.3
groq==0.8.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h5py==3.10.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.0
humansignal-drf-yasg==1.21.9
idna==3.4
ijson==3.3.0
imageio==2.31.6
importlib_metadata==8.4.0
inflection==0.5.1
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.25.1
ipython==8.12.2
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==3.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
karateclub==1.3.3
keras==2.15.0
kiwisolver==1.4.4
label-studio==1.13.1
label-studio-sdk==1.0.5
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.76
launchdarkly-server-sdk==8.2.1
lazy_loader==0.3
Levenshtein==0.26.0
libclang==16.0.6
lightning-utilities==0.9.0
llvmlite==0.41.1
lockfile==0.12.2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
multiprocess==0.70.15
mypy==1.6.1
mypy-extensions==1.0.0
namex==0.0.7
neo4j==5.20.0
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.9.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.33.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opentelemetry-api==1.27.0
opt-einsum==3.3.0
ordered-set==4.0.2
orjson==3.10.4
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==2.5.2
pluggy==1.3.0
ply==3.11
prompt-toolkit==3.0.36
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.0
psycopg2-binary==2.9.9
ptflops==0.7.1.1
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodomex==3.19.0
pydantic==2.7.3
pydantic_core==2.18.4
pyemd==0.5.1
Pygments==2.15.1
PyGSP==0.5.1
pyparsing==3.0.9
pyproject==1.3.1
pyRFC3339==1.1
pyrsistent==0.20.0
PyStanfordDependencies==0.3.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python-Levenshtein==0.26.0
python-louvain==0.16
pytorch-crf==0.7.2
pytorch-lightning==2.1.0
pytz==2022.7.1
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.0
qudida==0.0.4
rapidfuzz==3.9.7
redis==3.5.3
regex==2023.10.3
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==1.3.1
rich==13.7.0
rouge==1.0.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
semver==3.0.2
sentry-sdk==2.14.0
seqeval==1.2.2
setuptools==75.1.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
sqlparse==0.5.1
stack-data==0.2.0
sympy==1.12
tenacity==8.3.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0+cu118
torchinfo==1.8.0
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.16.0+cu118
tornado==6.3.2
tqdm==4.66.5
traitlets==5.7.1
transformers==4.34.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.38.4
wrapt==1.14.1
xmljson==0.2.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.20.2
pyemd==0.5.1
pytorch-transformers>=1.1.0
bllipparser
